# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Intelligence & Question Extraction Service",
    "description": "\n    ## Pragati Bharati Document Intelligence & Question Extraction Service\n    \n    A production-oriented, scalable microservice to ingest examination papers and question banks\n    in PDF (digital and scanned) and Image formats, converting them into structured, machine-readable\n    question sets.\n    \n    ### Key Features:\n    * **Asynchronous Document Ingestion**: Upload large documents without blocking clients.\n    * **Automated Question Segmentation**: Detect question numbers, bodies, and options.\n    * **Cross-Page Stitching**: Intelligently handles questions continuing across page breaks.\n    * **Answer Key Matching**: Pairs answer keys from inline sections or standalone documents.\n    * **Confidence & Review Pipeline**: Flags uncertain, partial, or low-quality extractions.\n    * **Security & Auth**: JWT-based bearer authentication and user authorization isolation.\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/documents/{document_id}/questions": {
      "get": {
        "tags": [
          "Questions & Answers"
        ],
        "summary": "Get Document Questions",
        "description": "Retrieve all extracted questions for a document with optional filters and pagination",
        "operationId": "get_document_questions_api_v1_documents__document_id__questions_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "confidence_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ConfidenceLevel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Confidence Filter"
            }
          },
          {
            "name": "question_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/QuestionType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Question Type"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/questions/{question_id}": {
      "get": {
        "tags": [
          "Questions & Answers"
        ],
        "summary": "Get Single Question",
        "description": "Retrieve full details of a specific question",
        "operationId": "get_single_question_api_v1_questions__question_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/answer-key": {
      "get": {
        "tags": [
          "Questions & Answers"
        ],
        "summary": "Get Document Answer Key",
        "description": "Retrieve extracted answer key entries for a document",
        "operationId": "get_document_answer_key_api_v1_documents__document_id__answer_key_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnswerKeyEntryResponse"
                  },
                  "title": "Response Get Document Answer Key Api V1 Documents  Document Id  Answer Key Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/export": {
      "get": {
        "tags": [
          "Questions & Answers"
        ],
        "summary": "Export Questions Structured",
        "description": "Returns questions in the clean, system-independent schema specified in\nProblem Statement Section 7.",
        "operationId": "export_questions_structured_api_v1_documents__document_id__export_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StructuredQuestionExport"
                  },
                  "title": "Response Export Questions Structured Api V1 Documents  Document Id  Export Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/review/documents/{document_id}/queue": {
      "get": {
        "tags": [
          "Review Queue & Validation"
        ],
        "summary": "Get Review Queue",
        "description": "Retrieve questions requiring human review:\n- Low confidence (<0.8)\n- Uncertain or unmatched answers\n- Cross-page continuation splits\n- Missing options or question numbers",
        "operationId": "get_review_queue_api_v1_review_documents__document_id__queue_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewQueueItemResponse"
                  },
                  "title": "Response Get Review Queue Api V1 Review Documents  Document Id  Queue Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/review/questions/{question_id}": {
      "put": {
        "tags": [
          "Review Queue & Validation"
        ],
        "summary": "Update Reviewed Question",
        "description": "Update and approve a reviewed question",
        "operationId": "update_reviewed_question_api_v1_review_questions__question_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewUpdatePayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnswerKeyEntryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "question_number": {
            "type": "integer",
            "title": "Question Number"
          },
          "correct_answer": {
            "type": "string",
            "title": "Correct Answer"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "source_page": {
            "type": "integer",
            "title": "Source Page"
          },
          "raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "question_number",
          "correct_answer",
          "source_page",
          "created_at"
        ],
        "title": "AnswerKeyEntryResponse"
      },
...
      "ConfidenceLevel": {
        "type": "string",
        "enum": [
          "CONFIDENT",
          "PARTIAL",
          "NEEDS_REVIEW"
        ],
        "title": "ConfidenceLevel"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OptionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "option_key": {
            "type": "string",
            "title": "Option Key"
          },
          "option_text": {
            "type": "string",
            "title": "Option Text"
          },
          "is_correct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Correct"
          }
        },
        "type": "object",
        "required": [
          "id",
          "option_key",
          "option_text"
        ],
        "title": "OptionResponse"
      },
      "QuestionListResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionResponse"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "total_count",
          "page",
          "page_size",
          "questions"
        ],
        "title": "QuestionListResponse"
      },
      "QuestionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "question_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Number"
          },
          "raw_number_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Number Label"
          },
          "question_text": {
            "type": "string",
            "title": "Question Text"
          },
          "question_type": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "source_pages": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Source Pages"
          },
          "confidence_score": {
            "type": "number",
            "title": "Confidence Score"
          },
          "confidence_level": {
            "$ref": "#/components/schemas/ConfidenceLevel"
          },
          "review_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Review Reasons"
          },
          "detected_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Answer"
          },
          "answer_explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Explanation"
          },
          "answer_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Source"
          },
          "has_images_or_tables": {
            "type": "boolean",
            "title": "Has Images Or Tables",
            "default": false
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/OptionResponse"
            },
            "type": "array",
            "title": "Options"
          },
          "metadata_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "question_text",
          "question_type",
          "confidence_score",
          "confidence_level",
          "created_at"
        ],
        "title": "QuestionResponse"
      },
      "QuestionType": {
        "type": "string",
        "enum": [
          "MULTIPLE_CHOICE",
          "TRUE_FALSE",
          "SHORT_ANSWER",
          "ESSAY",
          "NUMERICAL",
          "UNKNOWN"
        ],
        "title": "QuestionType"
      },
...
      "ReviewQueueItemResponse": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "question_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Number"
          },
          "question_text": {
            "type": "string",
            "title": "Question Text"
          },
          "confidence_score": {
            "type": "number",
            "title": "Confidence Score"
          },
          "confidence_level": {
            "$ref": "#/components/schemas/ConfidenceLevel"
          },
          "review_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Review Reasons"
          },
          "source_pages": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Source Pages"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "document_id",
          "question_text",
          "confidence_score",
          "confidence_level",
          "review_reasons",
          "source_pages"
        ],
        "title": "ReviewQueueItemResponse"
      },
      "ReviewUpdatePayload": {
        "properties": {
          "question_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Text"
          },
          "detected_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Answer"
          },
          "confidence_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceLevel"
              },
              {
                "type": "null"
              }
            ],
            "default": "CONFIDENT"
          },
          "clear_reasons": {
            "type": "boolean",
            "title": "Clear Reasons",
            "default": true
          }
        },
        "type": "object",
        "title": "ReviewUpdatePayload"
      },
      "StructuredQuestionExport": {
        "properties": {
          "question_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Number"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "source_pages": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Source Pages"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "confidence_level": {
            "type": "string",
            "title": "Confidence Level"
          },
          "review_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Review Reasons"
          }
        },
        "type": "object",
        "required": [
          "question",
          "confidence",
          "confidence_level"
        ],
        "title": "StructuredQuestionExport",
        "description": "Clean system-independent export format specified in Problem Statement Section 7"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}